Heart_Attack_Risk_Prediction/requirements.txt
Flask==3.0.0
joblib==1.3.2
numpy==1.24.3
pandas==2.0.3
twilio==8.10.0
scikit-learn==1.3.0
gunicorn==21.2.0
python-dotenv==1.0.0
Werkzeug==3.0.1
